+++ R/getfd.R
@@ -1,13 +1,10 @@
-#' getfd
-#'
+#' @title getfd
 #' @description Return normed path to current file directory
-#'
 #' @return Current file directory as string
-#' @examples
-#' getfd()
+#' @examples getfd()
 #' @export
 getfd <- function() {
-    if (interactive() && ((sys.nframe() > 0) && ("ofile" %in% ls(sys.frame(0))))) {
+    if (interactive() && sys.nframe() > 0 && "ofile" %in% ls(sys.frame(0))) {
         # probably started through `source` from interactive R session
         f <- sys.frame(1)$ofile
     } else {
@@ -18,8 +15,9 @@ getfd <- function() {
             filearg <- args[][[1]]
             f <- strsplit(filearg, "--file=")[[1]][2]
         } else {
-            f <- file.path(getwd(), "INTERACTIVE_SHELL")
+            stop("No file sourced. Maybe you're in an interactive shell?",
+                 call.=F)
         }
     }
-    return(normpath(dirname(f)))
+    return(normalizePath(dirname(f), winslash = "/"))
 }