+++ CHANGELOG.md
@@ -16,7 +16,14 @@ The format is loosely based on [Keep a Changelog](https://keepachangelog.com/en/
   - `Fixed` for bug fixes
   - `Security` in case of vulnerabilities
 
+## [0.0.0.9003] - 2021-11-26
+
+- `Changed`: `getfd` now throws an error if no sourced file can be found
+- `Changed`: `getprd` renamed to `getpd`
+- `Changed`: `getpd` now takes an argument of files used to decide which folder is the project root (default: `c(".git", "DESCRIPTION", "NAMESPACE")`)
+
 ## [0.0.0.9002] - 2021-11-18
+
 - `Added`: `where` command
 - `Added`: this changelog
 - `Added`: ``+`` operator